move examples class to sut (System Under test) package
