+++ services/src/test/java/io/scalecube/services/ErrorFlowTest.java
@@ -4,11 +4,11 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static reactor.core.publisher.Mono.from;
 
 import io.scalecube.services.api.ServiceMessage;
-import io.scalecube.services.examples.GreetingResponse;
-import io.scalecube.services.examples.GreetingServiceImpl;
 import io.scalecube.services.exceptions.BadRequestException;
 import io.scalecube.services.exceptions.ServiceUnavailableException;
 import io.scalecube.services.exceptions.UnauthorizedException;
+import io.scalecube.services.sut.GreetingResponse;
+import io.scalecube.services.sut.GreetingServiceImpl;
 
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
