+++ services/src/test/java/io/scalecube/services/ServiceCallLocalTest.java
@@ -16,10 +16,10 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
 import io.scalecube.services.api.ServiceMessage;
-import io.scalecube.services.examples.GreetingResponse;
-import io.scalecube.services.examples.GreetingServiceImpl;
 import io.scalecube.services.exceptions.ServiceException;
 import io.scalecube.services.routing.RoundRobinServiceRouter;
+import io.scalecube.services.sut.GreetingResponse;
+import io.scalecube.services.sut.GreetingServiceImpl;
 
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
