+++ services/src/test/java/io/scalecube/services/ServiceCallRemoteTest.java
@@ -17,9 +17,9 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import io.scalecube.services.api.ServiceMessage;
-import io.scalecube.services.examples.GreetingResponse;
-import io.scalecube.services.examples.GreetingServiceImpl;
 import io.scalecube.services.exceptions.ServiceException;
+import io.scalecube.services.sut.GreetingResponse;
+import io.scalecube.services.sut.GreetingServiceImpl;
 
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
