+++ services/src/test/java/io/scalecube/services/StreamingServiceTest.java
@@ -1,15 +1,14 @@
-package io.scalecube.services.streaming;
+package io.scalecube.services;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
-import io.scalecube.services.BaseTest;
-import io.scalecube.services.Microservices;
-import io.scalecube.services.ServiceCall;
 import io.scalecube.services.ServiceCall.Call;
 import io.scalecube.services.api.ServiceMessage;
+import io.scalecube.services.sut.QuoteService;
+import io.scalecube.services.sut.SimpleQuoteService;
 
 import com.codahale.metrics.MetricRegistry;
