+++ services/src/test/java/io/scalecube/services/routings/RoutersTest.java
@@ -13,16 +13,16 @@ import io.scalecube.services.Reflect;
 import io.scalecube.services.ServiceCall;
 import io.scalecube.services.ServiceCall.Call;
 import io.scalecube.services.api.ServiceMessage;
-import io.scalecube.services.examples.GreetingRequest;
-import io.scalecube.services.examples.GreetingResponse;
-import io.scalecube.services.examples.GreetingServiceImpl;
 import io.scalecube.services.routing.RandomServiceRouter;
 import io.scalecube.services.routing.Routers;
-import io.scalecube.services.routings.examples.CanaryService;
-import io.scalecube.services.routings.examples.DummyRouter;
-import io.scalecube.services.routings.examples.GreetingServiceImplA;
-import io.scalecube.services.routings.examples.GreetingServiceImplB;
-import io.scalecube.services.routings.examples.WeightedRandomRouter;
+import io.scalecube.services.routings.sut.CanaryService;
+import io.scalecube.services.routings.sut.DummyRouter;
+import io.scalecube.services.routings.sut.GreetingServiceImplA;
+import io.scalecube.services.routings.sut.GreetingServiceImplB;
+import io.scalecube.services.routings.sut.WeightedRandomRouter;
+import io.scalecube.services.sut.GreetingRequest;
+import io.scalecube.services.sut.GreetingResponse;
+import io.scalecube.services.sut.GreetingServiceImpl;
 
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
