+++ services/src/test/java/io/scalecube/services/routings/sut/CanaryService.java
@@ -1,9 +1,9 @@
-package io.scalecube.services.routings.examples;
+package io.scalecube.services.routings.sut;
 
 import io.scalecube.services.annotations.Service;
 import io.scalecube.services.annotations.ServiceMethod;
-import io.scalecube.services.examples.GreetingRequest;
-import io.scalecube.services.examples.GreetingResponse;
+import io.scalecube.services.sut.GreetingRequest;
+import io.scalecube.services.sut.GreetingResponse;
 
 import reactor.core.publisher.Mono;
